Pagination logic fixed.

+++ app/src/main/java/br/cericatto/leo/model/Search.kt
@@ -1,7 +1,7 @@
 package br.cericatto.leo.model
 
 data class Search(
-    val total_count: Long = 0L,
+    val total_count: Int = 0,
     val incomplete_results: Boolean = false,
     val items: List<Repo>
 )
